feat: add bearer token, delete refresh cookie, implement logic

+++ server/src/domains/userAuth/repositories/user.repository.ts
@@ -1,10 +1,10 @@
 import { prisma } from '../../../lib/prisma'
 
 export const authenticatedUserRepositoryFactory = () => {
-  const createToken = async (userId: string, accessToken: string) => {
+  const createAccessToken = async (userId: string, accessToken: string) => {
     const expiredAt = new Date()
     expiredAt.setDate(expiredAt.getDate() + 7)
-    console.log('token be', accessToken)
+
     return await prisma.token.create({
       data: {
         userId: userId,
@@ -14,19 +14,42 @@ export const authenticatedUserRepositoryFactory = () => {
     })
   }
 
+  /*const createRefreshToken = async (userId: string, refreshToken: string) => {
+    const expiredAt = new Date()
+    expiredAt.setDate(expiredAt.getDate() + 7)
+
+    return await prisma.token.create({
+      data: {
+        userId: userId,
+        token: refreshToken,
+        expiredAt,
+      },
+    })
+  }*/
+
   const findToken = async (userId: string, token: string) => {
     const foundToken = await prisma.token.findFirst({
-      where: { userId, token: token, expiredAt: { gte: new Date() } },
+      where: {
+        userId: userId,
+        token: token,
+
+        expiredAt: { gte: new Date() },
+      },
     })
 
     return foundToken || null
   }
-
   const deleteToken = async (userId: string, token: string) => {
-    await prisma.token.delete({
-      where: { token: token, userId },
+    const foundToken = await prisma.token.findFirst({
+      where: { userId, token },
     })
+
+    if (foundToken) {
+      await prisma.token.delete({
+        where: { id: foundToken.id },
+      })
+    }
   }
 
-  return { createToken, findToken, deleteToken }
+  return { createAccessToken, /*createRefreshToken*/ findToken, deleteToken }
 }
